requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pandas==2.2.3
numpy==2.1.3
faiss-cpu==1.9.0
runpod==1.7.5
pyarrow==18.1.0
fastparquet==2024.11.0
python-dotenv==1.0.1
pydantic==2.10.2
